Adding cloud for background
The background now contains a nice little cloud. Currently, the cloud is
simply repeated across the screen, but it will later be positioned
nicely. The source picture (translated by hand) is available as
cloud.png, which is being checked in.

The background color has been changed to a dark blue as well.